requirements.txt
google-generativeai
SpeechRecognition
pyaudio
streamlit
python-dotenv

-e .